style:optimize execute handler and handle db clusters
